requirements.txt
fastapi
uvicorn
httpx
langchain
openai
pinecone-client
pymupdf
python-dotenv